Add login persistence using SharedPreferences
This commit introduces a new feature that allows the application to remember
the login state of the user using SharedPreferences. The changes include adding
a new key 'isLogin' in the AppKeys class, updating the LoginCubit to set the
login state upon successful login, and modifying the SplashView to check the
login state and redirect accordingly. Additionally, a new LocalData class is
introduced to handle SharedPreferences operations, and the main.dart file is
updated to initialize SharedPreferences on app startup. The necessary
dependencies for SharedPreferences are also added to the pubspec.yaml and
pubspec.lock files, and the GeneratedPluginRegistrant.swift is updated to
register the SharedPreferences plugin for macOS.